Fix : le paramètre 'id' de la méthode doit maintenant être de type 'long' plutôt que de type 'int'.

+++ src/main/java/org/example/exception/TaskNotFoundException.java
@@ -2,7 +2,7 @@ package org.example.exception;
 
 public class TaskNotFoundException extends RuntimeException {
 
-    public TaskNotFoundException(int id) {
+    public TaskNotFoundException(long id) {
         super("La tâche d'id " + id + " n'existe pas");
     }
 }
